move get methods to utils
