+++ src/core/utils/transactions.ts
@@ -1,4 +1,5 @@
 import { FixedNumber } from '@ethersproject/bignumber';
+import { AddressZero } from '@ethersproject/constants';
 import {
   Provider,
   TransactionReceipt,
@@ -11,7 +12,11 @@ import { Address } from 'wagmi';
 
 import { i18n } from '../languages';
 import { createHttpClient } from '../network/internal/createHttpClient';
-import { SupportedCurrencyKey, smartContractMethods } from '../references';
+import {
+  ETH_ADDRESS,
+  SupportedCurrencyKey,
+  smartContractMethods,
+} from '../references';
 import {
   currentCurrencyStore,
   nonceStore,
@@ -32,6 +37,7 @@ import {
 
 import { parseAsset, parseUserAsset, parseUserAssetBalances } from './assets';
 import { getBlockExplorerHostForChain, isNativeAsset } from './chains';
+import { formatNumber } from './formatNumber';
 import { convertStringToHex } from './hex';
 import { capitalize } from './strings';
 
@@ -497,4 +503,81 @@ const TransactionOutTypes = [
 export const getDirection = (type: TransactionType) => {
   if (TransactionOutTypes.includes(type)) return 'out';
   return 'in';
+};
+
+export const getExchangeRate = ({ type, changes }: RainbowTransaction) => {
+  if (type !== 'swap') return;
+
+  const tokenIn = changes?.filter((c) => c?.direction === 'in')[0]?.asset;
+  const tokenOut = changes?.filter((c) => c?.direction === 'out')[0]?.asset;
+
+  const amountIn = tokenIn?.balance.amount;
+  const amountOut = tokenOut?.balance.amount;
+  if (!amountIn || !amountOut) return;
+
+  const fixedAmountIn = FixedNumber.fromString(amountIn);
+  const fixedAmountOut = FixedNumber.fromString(amountOut);
+
+  const rate = fixedAmountOut.divUnsafe(fixedAmountIn).toString();
+  if (!rate) return;
+
+  return `1 ${tokenIn.symbol} ≈ ${formatNumber(rate)} ${tokenOut.symbol}`;
+};
+
+export const getAdditionalDetails = (transaction: RainbowTransaction) => {
+  const exchangeRate = getExchangeRate(transaction);
+  const { asset, changes, approvalAmount, contract, type } = transaction;
+  const nft = changes?.find((c) => c?.asset.type === 'nft')?.asset;
+  const collection = nft?.symbol;
+  const standard = nft?.standard;
+  const tokenContract =
+    asset?.address !== ETH_ADDRESS && asset?.address !== AddressZero
+      ? asset?.address
+      : undefined;
+
+  const tokenAmount =
+    !nft && !exchangeRate && tokenContract
+      ? changes?.find((c) => c?.asset.address === tokenContract)?.asset.balance
+          .amount
+      : undefined;
+
+  const approval = type === 'approve' &&
+    approvalAmount && {
+      value: approvalAmount,
+      label: getApprovalLabel(transaction),
+    };
+
+  if (
+    !tokenAmount &&
+    !tokenContract &&
+    !exchangeRate &&
+    !collection &&
+    !standard &&
+    !approval
+  )
+    return;
+
+  return {
+    asset,
+    tokenAmount: tokenAmount && `${formatNumber(tokenAmount)} ${asset?.symbol}`,
+    tokenContract,
+    contract,
+    exchangeRate,
+    collection,
+    standard,
+    approval,
+  };
+};
+
+export const getApprovalLabel = ({
+  approvalAmount,
+  asset,
+  type,
+}: Pick<RainbowTransaction, 'type' | 'asset' | 'approvalAmount'>) => {
+  if (!approvalAmount || !asset) return;
+  if (approvalAmount === 'UNLIMITED') return i18n.t('approvals.unlimited');
+  if (type === 'revoke') return i18n.t('approvals.no_allowance');
+  return `${formatNumber(formatUnits(approvalAmount, asset.decimals))} ${
+    asset.symbol
+  }`;
 };
